Move inner classes to outer level, to avoid too big class files
